pyproject.toml
[tool.poetry]
name = "langchain-demo"
version = "0.1.0"
description = ""
authors = ["marvlach <marios.vlachog@gmail.com>"]
readme = "README.md"
packages = [{include = "langchain_demo"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "1.0.0"
langchain = "0.0.136"
pinecone-client = "2.2.1"
openai = "^0.27.7"
tiktoken = "^0.4.0"
pymssql = "^2.2.7"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
